fake_news_model/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==1.5.3
tensorflow==2.11.0
requests~=2.28.2
joblib~=1.2.0
contractions~=0.1.73
beautifulsoup4~=4.11.2
validators~=0.20.0